+++ packages/braid/vitest.config.ts
@@ -1,7 +1,8 @@
-import { defineConfig } from "vitest/config";
+import { configDefaults, defineConfig } from "vitest/config";
 
 export default defineConfig({
 	test: {
+		reporters: ["verbose", ...configDefaults.reporters],
 		environment: "node",
 		globals: true,
 		include: ["src/**/*.spec.ts"],